requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.3.7
    # via aiohttp
aiohttp==3.10.4
    # via langchain
aiosignal==1.3.1
    # via aiohttp
alembic==1.13.2
    # via optuna
altair==5.4.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   aiohttp
    #   cmd2
    #   jsonschema
    #   referencing
autopage==0.5.2
    # via cliff
blinker==1.8.2
    # via streamlit
build==1.2.1
    # via pip-tools
cachetools==5.5.0
    # via streamlit
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pip-tools
    #   streamlit
cliff==4.7.0
    # via optuna
cmaes==0.11.1
    # via optuna
cmd2==2.4.3
    # via cliff
colorama==0.4.6
    # via
    #   build
    #   click
    #   colorlog
    #   loguru
    #   tqdm
colorlog==6.8.2
    # via optuna
distro==1.9.0
    # via openai
fastjsonschema==2.20.0
    # via nbformat
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
hydra-core==1.3.2
    # via
    #   -r requirements.in
    #   hydra-optuna-sweeper
hydra-optuna-sweeper==1.2.0
    # via -r requirements.in
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via
    #   altair
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-core==5.7.2
    # via nbformat
langchain==0.2.14
    # via -r requirements.in
langchain-core==0.2.33
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.22
    # via -r requirements.in
langchain-text-splitters==0.2.2
    # via
    #   -r requirements.in
    #   langchain
langsmith==0.1.99
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
loguru==0.7.2
    # via -r requirements.in
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
narwhals==1.4.2
    # via altair
nbformat==5.10.4
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   cmaes
    #   langchain
    #   optuna
    #   pandas
    #   pyarrow
    #   pydeck
    #   scipy
    #   streamlit
omegaconf==2.3.0
    # via hydra-core
openai==1.41.0
    # via
    #   -r requirements.in
    #   langchain-openai
optuna==2.10.1
    # via hydra-optuna-sweeper
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   altair
    #   build
    #   hydra-core
    #   langchain-core
    #   optuna
    #   streamlit
pandas==2.2.2
    # via
    #   -r requirements.in
    #   streamlit
pbr==6.0.0
    # via stevedore
pillow==10.4.0
    # via streamlit
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via jupyter-core
prettytable==3.11.0
    # via cliff
protobuf==5.27.3
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.8.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyperclip==1.9.0
    # via cmd2
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pyreadline3==3.4.1
    # via cmd2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
python-gitlab==4.9.0
    # via -r requirements.in
pytz==2024.1
    # via pandas
pywin32==306; sys_platform == 'win32'
    # via jupyter-core
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   cliff
    #   langchain
    #   langchain-core
    #   omegaconf
    #   optuna
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langsmith
    #   python-gitlab
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via python-gitlab
rich==13.7.1
    # via streamlit
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
scipy==1.14.0
    # via optuna
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.32
    # via
    #   alembic
    #   langchain
    #   optuna
stevedore==5.2.0
    # via cliff
streamlit==1.37.1
    # via -r requirements.in
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-core
    #   streamlit
tiktoken==0.7.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.4.1
    # via streamlit
tqdm==4.66.5
    # via
    #   openai
    #   optuna
traitlets==5.14.3
    # via
    #   jupyter-core
    #   nbformat
typing-extensions==4.12.2
    # via
    #   alembic
    #   altair
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.2
    # via streamlit
wcwidth==0.2.13
    # via
    #   cmd2
    #   prettytable
wheel==0.44.0
    # via pip-tools
win32-setctime==1.1.0
    # via loguru
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
